+++ iftttie/utils.py
@@ -6,14 +6,16 @@ from typing import AsyncIterable, TypeVar
 
 import click
 from aiohttp import StreamReader
-from argon2 import PasswordHasher
 from click import echo, style
+from passlib.handlers.pbkdf2 import pbkdf2_sha256
 
 T = TypeVar('T')
 
 
 async def read_lines(reader: StreamReader) -> AsyncIterable[str]:
-    """Provides async iterable interface for a reader."""
+    """
+    Provides async iterable interface for a reader.
+    """
     while True:
         line: bytes = await reader.readline()
         if not line:
@@ -22,13 +24,17 @@ async def read_lines(reader: StreamReader) -> AsyncIterable[str]:
 
 
 async def iterate_queue(queue: Queue[T]) -> AsyncIterable[T]:
-    """Provides async iterable interface for a queue."""
+    """
+    Provides async iterable interface for a queue.
+    """
     while True:
         yield await queue.get()
 
 
 def import_from_string(name: str, code: str) -> ModuleType:
-    """Imports module from code passed as the parameter."""
+    """
+    Imports module from code passed as the parameter.
+    """
     module = ModuleType(name)
     module.__package__ = 'iftttie'
     exec(code, module.__dict__)
@@ -49,17 +55,13 @@ def hash_password(password: str):
     """
     Print password hash for the IFTTTie dashboard authentication.
     """
-    hash_ = hasher.hash(password)
+    hash_ = pbkdf2_sha256.hash(password)
 
     echo()
     echo(style('Success! Put the following to your configuration module:', fg='green'))
     echo()
     echo(style(f"USERS = [\n    ('username', '{hash_}'),\n]", fg="yellow"))
     echo()
-
-
-# Parameters adjusted for Raspberry Pi Zero W.
-hasher = PasswordHasher(time_cost=1, memory_cost=8, parallelism=1)
 
 
 if __name__ == '__main__':